# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "campus-life-platform",
    "description": "校园生活平台（Python 重写）模块化单体后端",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/courses": {
      "get": {
        "tags": [
          "course"
        ],
        "summary": "List All",
        "operationId": "list_all_api_courses_get",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Keyword"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "course"
        ],
        "summary": "Create",
        "operationId": "create_api_courses_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CourseOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}": {
      "get": {
        "tags": [
          "course"
        ],
        "summary": "Detail",
        "operationId": "detail_api_courses__course_id__get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/reviews": {
      "post": {
        "tags": [
          "course"
        ],
        "summary": "Review",
        "operationId": "review_api_courses__course_id__reviews_post",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseReviewCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CourseReviewOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/canteens": {
      "get": {
        "tags": [
          "canteen"
        ],
        "summary": "List All",
        "operationId": "list_all_api_canteens_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "canteen"
        ],
        "summary": "Create",
        "operationId": "create_api_canteens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CanteenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CanteenOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/canteens/stalls": {
      "post": {
        "tags": [
          "canteen"
        ],
        "summary": "Create Stall Endpoint",
        "operationId": "create_stall_endpoint_api_canteens_stalls_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StallCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_StallOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/canteens/dishes": {
      "post": {
        "tags": [
          "canteen"
        ],
        "summary": "Create Dish Endpoint",
        "operationId": "create_dish_endpoint_api_canteens_dishes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DishOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/canteens/dishes/{dish_id}": {
      "get": {
        "tags": [
          "canteen"
        ],
        "summary": "Dish Detail",
        "operationId": "dish_detail_api_canteens_dishes__dish_id__get",
        "parameters": [
          {
            "name": "dish_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dish Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/canteens/dishes/{dish_id}/reviews": {
      "post": {
        "tags": [
          "canteen"
        ],
        "summary": "Review",
        "operationId": "review_api_canteens_dishes__dish_id__reviews_post",
        "parameters": [
          {
            "name": "dish_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dish Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CanteenReviewCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CanteenReviewOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_CanteenOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CanteenOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[CanteenOut]"
      },
      "ApiResponse_CanteenReviewOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CanteenReviewOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[CanteenReviewOut]"
      },
      "ApiResponse_CourseOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CourseOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[CourseOut]"
      },
      "ApiResponse_CourseReviewOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CourseReviewOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[CourseReviewOut]"
      },
      "ApiResponse_DishOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DishOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[DishOut]"
      },
...
      "ApiResponse_StallOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StallOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[StallOut]"
      },
...
      "ApiResponse_dict_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse[dict]"
      },
      "ApiResponse_list_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse[list]"
      },
...
      "CanteenCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "location": {
            "type": "string",
            "title": "Location",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CanteenCreate"
      },
      "CanteenOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "stalls": {
            "items": {
              "$ref": "#/components/schemas/StallOut"
            },
            "type": "array",
            "title": "Stalls",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "location"
        ],
        "title": "CanteenOut"
      },
      "CanteenReviewCreate": {
        "properties": {
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "default": 5
          },
          "content": {
            "type": "string",
            "title": "Content",
            "default": ""
          }
        },
        "type": "object",
        "title": "CanteenReviewCreate"
      },
      "CanteenReviewOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "dish_id": {
            "type": "string",
            "title": "Dish Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dish_id",
          "user_id",
          "rating",
          "content"
        ],
        "title": "CanteenReviewOut"
      },
      "CourseCreate": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "teacher": {
            "type": "string",
            "title": "Teacher",
            "default": ""
          },
          "credits": {
            "type": "integer",
            "title": "Credits",
            "default": 0
          },
          "semester": {
            "type": "string",
            "title": "Semester",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "CourseCreate"
      },
      "CourseOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "teacher": {
            "type": "string",
            "title": "Teacher"
          },
          "credits": {
            "type": "integer",
            "title": "Credits"
          },
          "semester": {
            "type": "string",
            "title": "Semester"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "teacher",
          "credits",
          "semester"
        ],
        "title": "CourseOut"
      },
      "CourseReviewCreate": {
        "properties": {
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "default": 5
          },
          "content": {
            "type": "string",
            "title": "Content",
            "default": ""
          }
        },
        "type": "object",
        "title": "CourseReviewCreate"
      },
      "CourseReviewOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "user_id",
          "rating",
          "content"
        ],
        "title": "CourseReviewOut"
      },
      "DishCreate": {
        "properties": {
          "stall_id": {
            "type": "string",
            "title": "Stall Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "price": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Price",
            "description": "单位：分",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "stall_id",
          "name"
        ],
        "title": "DishCreate"
      },
      "DishOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "stall_id": {
            "type": "string",
            "title": "Stall Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "price": {
            "type": "integer",
            "title": "Price"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stall_id",
          "name",
          "price"
        ],
        "title": "DishOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StallCreate": {
        "properties": {
          "canteen_id": {
            "type": "string",
            "title": "Canteen Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "canteen_id",
          "name"
        ],
        "title": "StallCreate"
      },
      "StallOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "canteen_id": {
            "type": "string",
            "title": "Canteen Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "dishes": {
            "items": {
              "$ref": "#/components/schemas/DishOut"
            },
            "type": "array",
            "title": "Dishes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "canteen_id",
          "name"
        ],
        "title": "StallOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}